+++ opero/opero/doctype/wash_expertise_area/wash_expertise_area.py
@@ -1,5 +1,5 @@
 from frappe.model.document import Document
 
 
-class WaSHExpertiseArea(Document):
+class WASHExpertiseArea(Document):
 	pass